Add a command to submit patch series to upstream Git
Upstream Git wants to have code contributions as mails. This gets a
little tedious when the contributions have to be re-worked and
re-submitted.

Let's have a helper that makes that substantially less tedious.

To avoid the problem noticed by Brian M. Carlson where dash would
transmogrify the '\n' into a newline in

	x="$(echo "printf '\\n'")"
	echo "$x"

we force the script to be run through Bash.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>