mail-patch-series: use the `upstream` remote
While this developer used the name `junio` for a long time, it is easier
to call all upstream remotes in all projects `upstream` rather than
having them be named after the maintainer.

An exception is Cygwin, as this developer's scenario is such that
upstream is actually MSYS2, and "upstream's upstream" is Cygwin, so
let's keep that name.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>