cli: fix tag message parsing
The tag message was not matched as a multi-line message, and when the
Message-IDs were parsed, they were not shifted into the in_reply_to
array (but rather the first element, if any, was removed). Also, the
in_reply_to array was populated with the command-line options intended
for `git format-patch` rather than the bare Message-IDs.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>