Extract proper classes from the script
This commit disentangles the command-line script to come up with some
idea of Object-Orientation.

With this change, we also extract the console log messages into a
central method that orchestrates the other methods. That will hopefully
make it easier to reuse the code as a library in the future.

We use this opportunity to switch from a line-oriented editing process
to a regex-based process, which should not only be more efficient, but
also help memory consumption and robustness in the long run.

Next stop: unit testing.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>