git: add a wrapper for `git rev-parse --verify <argument>`
To make it more useful, return `undefined` if the command failed to
verify the argument.

Note that it will *always* return the argument back to the user if it is
a hex string of length 40. This is consistent with `rev-parse`. To
verify objects by full SHA-1, you have to add `^{blob}` or similar.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>