ProjectOptions: extract initialization intended for cli usage
In the command-line invocation, we were sure that the script was
launched in a worktree.

However, in preparation for GitGitGadget (which is a GitHub App, at
least partially), we will want to have a way where we do *not* store the
configuration in tags, and do not work on local branches, but on Pull
Requests instead.

So let's rename the function intended to initialize the ProjectOptions
instance for command-line usage to `getLocal()`, and generalize the
`get()` function to allow for passing in information that is no longer,
say, a local branch description (the cover letter in command-line usage,
but for GitGitGadget that will be the Pull Request's description instead).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>