Start to test the main functionality of GitGitGadget
With this added test, we can now start to gain confidence that
GitGitGadget does what it is supposed to do.

And even better: if it should not to what it is supposed to do at some
stage, we can easily teach this test to expect what we want, and then
fix the code to actually do it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>