Let's use `git range-diff` directly
The functionality provided by tbdiff (https://github.com/rast/tbdiff)
was used in the command-line version of GitGitGadget, but we cannot use
tbdiff in GitGitGadget because that web app will run in Azure, in the
free tier, on Windows, and tbdiff (like any advanced Python script using
numpy) is notoriously hard to get to run on Windows.

Happily, the `range-diff` command (which was originally contributed as
`branch-diff` but renamed due to popular opinion) is on its way into
git.git, and we can easily use a custom MinGit that provides
`range-diff` out of the box.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>