Start the Web App
This is the real thing. This adds the Probot-based web app that serves
commands issued via comments on Pull Requests.

The main UI is the `/submit` command: simply add a comment with the bare
command `/submit` to a PR at https://github.com/gitgitgadget/git, and
the web app will pick it up, send the patch series, and reply with a
link to the sent cover letter.

A few bits are still left in the TODO.md document, we will get there in
a bit, though.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>