GitGitGadget test: verify that iteration > 1 is handled correctly
One of the trickier things is to send v2 and later of any given patch
series, as that often means that somebody reviewed the code and it would
be nice to give reviewers a break by not forcing them to review
everything again (and to figure out what has changed, actually).

To make all this a bit easier on reviewers, we want to make the cover
letter a reply to the cover letter of the previous iteration, and
include a range-diff.

And to make that possible, we need to make sure that the Git notes are
being updated accordingly.

Let's make sure that all of that works correctly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>