Web App: fetch the notes/pull refs before submitting
This is needed to make sure that we start from the correct state.

Note: this will fail if a user has been added to the allowed users
outside of the web app, because it will not re-fetch and parse the notes
before testing permissions. This is okay, though, as the primary way to
allow new users will be through the `/allow` and `/disallow` commands
that are still to be implemented. And those commands will run in the web
app, and update the options accordingly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>