PatchSeries: force-add mail metadata notes
When updating the metadata corresponding to a certain commit, say, when
it is sent *again* as part of a subsequent iteration, there is already a
Git note, which we want to overwrite. This can happen when `master` does
not advance and e.g. only the last commit of many had to be changed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>