Avoid a specific Git version in format-patch's signature
The `git format-patch` command has the (sometimes annoying) habit of
adding a "signature" that reflects the very exact Git version used to
generate the patch series.

That might have made sense when `git format-patch` was in its infancy
and changed its output subtly from time to time.

But it makes testing next to impossible, so let's just override that
signature with our own.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>