Fix bug where only last Cc: would be used
The code assumed that the Cc: list would be in one single Cc: line as a
comma-separated list. That is, however, not what `git format-patch`
does: it generates separate Cc: lines.

This issue is supposedly fixed in Git v2.18.0, but in dugite we still
use v2.17.0. In any case, we should heed Postel's Law and be lenient in
our inputs.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>