Add a class to use the mail<->commit mapping
One part of GitGitGadget is an automated build that maintains the
`commit-to-mail` and `mail-to-commit` notes. These notes can be used to
try to figure out which upstream commits ("git.git commits") correspond
to which mails on the Git mailing list, and vice versa.

We need such a mapping to determine efficiently when/if any GitGitGadget
PR has been integrated into any of the four integration branches of
git.git yet.

For now, this class only has code to use `refs/notes/mail-to-commit` of
https://github.com/gitgitgadget/git.

Later commits will add code to actually use it.

Side note: In https://github.com/gitster/git, Junio Hamano publishes a
notes ref under the name `amlog` that maps some commits to their
corresponding Message-IDs, and some of them are even accurate. Too many
are not, and too many are missing, so we cannot use them here.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>