Add a class to annotate/change the Pull Requests on GitHub
The motivation behind this class is that we want to install an automated
"Continuous Integration" job to monitor e.g. the `mail-to-commit` notes
(see the previous commits for details on that one) and to update the PR
on GitHub accordingly.

For example, once the patches have been integrated into the `pu` branch,
we want to annotate the PR's commits with links to the corresponding
commits in the `pu` branch, mark the PR as being integrated into `pu`, etc.

We use the Checks API here instead of the Commit Statuses API, despite
the user experience of the latter being a bit nicer, because Commit
Statuses can never be removed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>