GitGitGadget: require parameter pointing to the "envelope directory"
GitGitGadget is configured via Git config settings, e.g.
gitgitgadget.workDir, which tells GitGitGadget where the worktree with
the git.git commits and the notes ref is to be found.

To be able to test this, we need to have a way *not* to use the Git
config of GitGitGadget's source code. So we add a parameter to make it
explicit where GitGitGadget should look for its configuration.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>