Hotfix: pass along the vouching user for /allow and /disallow
The `allowUser()` and `denyUser()` methods of the `GitGitGadget` class
require *two* parameters, the vouching user and the user to be added
(or removed) from the list of allowed users.

By mistake, the callers in `index.js` only passed the latter, which is
fixed by this commit.

It would probably be a good idea to convert `index.js` to Typescript,
too, to avoid such stupid mistakes.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>