Add a command-line helper to inspect/update stuff
There are already a couple of open Pull Requests at
https://github.com/gitgitgadget/git/pulls, and we will want to update
the metadata (whose "schema" we just expanded) to reflect which PRs are
open, and which mails are "active" (read: contain the patches
corresponding to open PRs).

So we need a tool to fix things up after the fact. This is it. This tool
will grow all the functionality we need in that respect.

Signed-off-by: Johannes Schindelin <johannes.Schindelin@gmx.de>