misc-helper: optionally find out the upstream commit for an original one
When going through that lossy medium called email, the original commit
SHA-1 gets lost, and therefore it can be daunting to figure out what the
heck happened to ones commits on the way into git.git.

When using GitGitGadget, we record the Message-ID corresponding to the
mail that was generated and sent from the original commits, and the Git
maintainer thankfully maintains a similar log (from Message-ID to commit
in `pu`).

This commit adds the new sub-command `lookup-upstream-commit` to the
`misc-helper` script, to automate looking all this up.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>