CIHelper: add a method to auto-update the mail-commit-mapping ref
Whenever the CIHelper wants to do anything with the MailCommitMapping
class, it makes sense to update the notes ref and the branches.

We only need to do that at most once per CIHelper instantiation, though
(the idea being that the CIHelper will be instantiated via automated
builds when any of the upstream branches or the `amlog` notes ref
changes).

So let's add a method that we can call to ensure that those refs were
updated.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>