CIHelper: find upstream equivalents for PRs' commits
When working with Git's code contribution process, it is often necessary
to get quickly from the original commit to the corresponding upstream
commit (because patches are applied from mails, so the commit name is
necessarily different from the original).

This commit adds a method by which GitGitGadget can use the
`gitgitgadget` and `mail-to-commit` notes to identify the upstream
commit, given an original commit.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>