GitNotes: add a method to get the last commit note
Commit notes are appended using the `git notes append` command, which
means that either a new commit note is created, or the contents of an
existing one are combined with an empty line and then the new note.

In our use case, we can safely assume that the notes we append will not
contain empty lines, so an easy way to get at the latest note that was
appended is to simply get the commit notes and strip everything until
the last empty line (if any).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>