Avoid duplicate Cc: header lines
We already have code that tries to reuse existing Cc: header lines
instead of adding a new one.

However, that code failed to anticipate the situation when the `Cc:`
header is the last one in the header block, as it always expected
another line to come after that.

Let's just handle that situation correctly.

This fixes https://github.com/gitgitgadget/gitgitgadget/issues/16#L271

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>