PatchSeries: use merge base for the `--base` argument
The `git format-patch` command is really unhappy when we specify a
`--base` argument that is not an ancestor of the processed commits.

And while there is support for `--base=auto`, that option is absolutely
useless for our use case, as it inserts the merge base of the *current*
branch and *its upstream*, both of which are of course nonsensical in
GitGitGadget's case.

So let's just spawn yet another process to get the correct merge base.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>