patch-series: override Date: headers with current date
We are trying to emulate `git send-email`, and that command ignores the
date that was created by `git format-patch` in favor of the current
time.

To be precise, we assign incrementing dates to the consecutive mail, one
second apart, until the last mail, which has the current date.

This helps users keep their inbox more organized, especially when they
do not thread their messages.

Original-patch-by: Derrick Stolee <dstolee@microsoft.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>