Work harder to avoid duplicate Cc: header lines
We already have code that tries to reuse existing Cc: header lines
instead of adding a new one.

And we already fixed the situation where our code failed to anticipate the situation when the `Cc:` header is the last one in the header block.

But we still failed to handle the case correctly where the existing Cc:
header was already a multi-line one.

Let's also handle that case.

This hopefully fixes #16 for good.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>