lookup-commit.sh: a helper providing a commit -> Message-ID mapping
The original idea for keeping GitGitGadget's PRs up to date, despite the
fact that commits will *never* make it into git.git unchanged (they will
all be sent via mail, then applied with a complete different commit
hash), was to use the `amlog` notes that used to provide a mapping from
original Message-ID to commit in git.git.

During the exploration/development of this part of GitGitGadget, serious
gaps and mistakes were discovered in the `amlog` notes, and rather than
fixing it, the Git maintainer chose to take away the reverse mapping,
and not fix the gaps at all.

As a consequence, this developer was forced to re-think the design of
this feature. Going back to another exploration contributed as
https://public-inbox.org/git/nycvar.QRO.7.76.6.1804041821420.55@ZVAVAG-6OXH6DA.rhebcr.pbec.zvpebfbsg.pbz/
in the wake of GitMerge 2017, this commit adds a script that takes
commits from git.git and tries to match them via their timestamp to the
corresponding Message-IDs in https://public-inbox.org/git.

As this is not fool-proof, quite a few mappings were added by hand,
others were ignored, and all unmatched commits from before 2012 were
left as-are, to be fixed later, if necessary.

The script now also has an `update` mode which updates the
`refs/notes/commit-to-mail` notes in https://github.com/gitgitgadget/git
(which are essentially what `amlog` should have been).

The new idea is to install a VSTS job that keeps this notes ref up to
date with `pu`.

For convenience (and also to accommodate the surprising fact that pretty
much all Git oldtimes flatly refuse to touch anything that is
Typescript, claiming that they are unfamiliar with it, despite the fact
that it is *really* close to Javascript, but whatever), this script is a
Unix shell script.

Another script will be added later that will be installed in the same
VSTS job to try to match the new commits to GitGitGadget's PRs' commits,
and then update the respective PRs' statuses.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>