Format cover letters
There was already a Markdown renderer that this developer added
specifically to process cover letters, but it was forgotten.

This not only renders Markdown a bit nicer, but it also wraps the lines
at 76 columns, in line with what is common on the Git mailing list.

This fixes https://github.com/gitgitgadget/gitgitgadget/issues/26

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>