misc-helper: allow continuing non-GitGitGadget submissions
It is pretty nice to use GitGitGadget. But what if a patch series had
already been started without GitGitGadget?

Fear not. This script can be used to initialize the PR metadata
accordingly.

It is still a very manual, error-prone process (the information about
the number of patches should be taken from the cover letter, the tip
commit should be verified against the latest diff, and the base commit
should be determined automatically, just like the iteration).

If need be, we can start from here, automating one little piece by one
little piece.

The primary reason for this was to allow submitting v2 iterations for
rebase-in-c-2-basic and rebase-in-c-4-opts:
https://github.com/gitgitgadget/git/pull/32 and
https://github.com/gitgitgadget/git/pull/33

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>