package-lock: run npm audit
GitGitGadget should not be vulnerable at all, because the UI is limited
to a precious small set of commands that are automatically rejected
unless issued by users in the permitted list.

But, you know, as they say: better safe 'n sorry.

This commit is the output of a straight run of `npm audit fix`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>