misc-helper: start implementing the Pipeline side of the new strategy
The "new strategy" being to trigger an Azure Pipeline via an Azure
Function (instead of doing everything in one opaque WebApp).

To that end, this commit ports the Javascript code in `index.js` to the
`misc-helper` (catching one bug: the `/disallow` command would have
called the `disallowUser()` method that does not exist; It is called
`denyUser()` instead).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>