Add an Azure Function to back the GitGitGadget GitHub App
Earlier, we used an Azure WebApp. The free tiers do not allow these
WebApps to be "always on", which resulted in roughly 2 minutes of
waiting time before the WebApp would kick in, and there was no way for
users to see the progress. Besides, two concurrent `/submit` commands
would wreak havoc because they used the same Git worktree.

Let's use an Azure Function instead that does nothing else but kick off
an Azure Pipeline build. Solving all of the above-mentioned problems.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>