ci: greet new GitGitGadget users
The UI of GitGitGadget is hardly intuitive for the moment: you have to
*know* what to do with those Pull Requests, and you have to know that
you have to get on that list of permitted users, and you have to know
how you can get there.

All things new users really are unlikely to know.

So let's tell them: when they open a Pull Request, say hi, explain what GitGitGadget is, how it works, and how they can use it.

This task is of course intended to be run in an Azure Pipeline,
triggered by the very Pull Request where the comment should be added.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>