Add a helper class to mirrors replies on the Git list back to the PRs
It is all nice and dandy to submit patch series from PRs, but you cannot
really see the reviewers' comments in the PRs, making the entire
exercise a little less fun.

Let's do something about it.

This new class only helps with processing the mails, but it is not
hooked up with any command-line interface yet, and it also does not
store the state correctly yet, e.g. "what was the latest mail we
processed?" (to allow for efficient incremental processing).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>