git: optionally accept a handler to process stdout line by line
With this option, we allow callers to specify a callback that handles
every line of the input individually.

This will come in handy when we will start processing the output of
`git log -p` in public-inbox/git, as we will be able to handle each
commit individually, together with its diff, from which we can extract
the corresponding mbox.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>