CIHelper: prepare for updating a commit mapping explicitly
It is nice and dandy if the commit-to-mail and mail-to-commit notes are
reliable, but when they are not (and they seem not to, for some reason
there was an uptick of unmatched commits recently), it becomes less nice
and less dandy.

In preparation for spending a bit more effort to identify upstream
commits, allow specifying the upstream commit explicitly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>