updateCommitMappings(): skip Git GUI PRs
It does not make sense to look for branch names and/or commits in
`upstream/pu` for these PRs, as the Git maintainer does not pick up Git
GUI patches directly: the Git GUI maintainer does, and they work in a
different public repository.

This should make the automated task at
https://dev.azure.com/gitgitgadget/git/_build?definitionId=2&_a=summary
work again (which is currently broken due to an out-of-memory when
trying to match Git GUI commits to `upstream/pu` via `git range-diff`).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>