For Git GUI contributions, Cc: the appropriate maintainer
Only Git contributions should be Cc:ed to Junio Hamano, the Git GUI
contributions should be Cc:ed to the Git GUI maintainer, Pratyush Yadav.

To detect whether the Pull Request that is about to be sent is a Git GUI
one, we use the fact that `git-gui.sh` lives in `git-gui/` in git.git's
branches, but in the top-level directory in Git GUI's branches.

Concretely, we test for the presence of the root commit of Git GUI's
commit history, and we test whether the target branch has `git-gui.sh`
in its top-level directory. If both tests pass, this is a Git GUI Pull
Request, and it needs to be Cc:ed to the Git GUI maintainer instead of
the Git maintainer.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>