Make the `range-diff`s more useful
In GitGitGadget, we can be _pretty_ certain that subsequent patch series
iterations' patches are correlated to previous ones, even if they might
be near rewrites. But the `git range-diff` command used by GitGitGadget
is relatively conservative when it comes to identifying matching
patches: it avoids mistaking dropped/added patches for different
versions of the same patch.

Let's override the default "creation factor", i.e. try harder to find
matching patches, as suggested by the Git maintainer, Junio Hamano, in
https://public-inbox.org/git/xmqqimp26808.fsf@gitster-ct.c.googlers.com/

The default creation factor is 60 (roughly speaking, it wants 60% of the
lines to match between two patches, otherwise it considers the patches
to be unrelated).

So which creation factor to use?

In the patch series iteration that triggered Junio's suggestion, a
creation factor of 67 would have been sufficient to identify the the
matching patches correctly.

In a more difficult case, however, where a patch was rewritten almost
completely, and only parts of the commit message survived the
development, it would have taken a creation of 93:
https://public-inbox.org/git/pull.361.v3.git.gitgitgadget@gmail.com/

So let's be conservative in the other direction in GitGitGadget's case,
and use a creation factor of 95. That might still result in false
negatives (or false positives!), but it is likely a lot better than 60.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>