Support additional nodemailer options
For testing, an alternate SMTP server may be used to avoid sending
emails.  Setting "gitgitgadget.smtpOpts" in the git config will
allow other nodemailer options to be specified.

Ethereal.email is an example server which does not actually send the
emails.  It does require other options to be passed to nodemailer.

Sample setting:
    await worktree.git(["config",
        '--add', "gitgitgadget.smtpHost", "smtp.ethereal.email"]);
    await worktree.git(["config",
        '--add', "gitgitgadget.smtpOpts",
        '{ "port": 587, "secure": false,
		"tls": { "rejectUnauthorized": false } }']);

[jes: change tsconfig.json to support `Object.enries()` as described in
https://patrickdesjardins.com/blog/typescript-and-object-entries]