Azure Function: allow being triggered from git/git
This is the most important step to let GitGitGadget support PRs at
https://github.com/git/git: do not reject webhooks originating from
there, and trigger a different pipeline than for events being sent over
from gitgitgadget/git.

The different Azure Pipeline is important because the pipeline has to be
registered as a CI build for the respective repository, so that it shows
up in the "Checks" section of the Pull Requests in question.

While at it, let's document the relationship between the (single) Azure
Function and the (multiple) Azure Pipelines that need to work together
to implement the backend of GitGitGadget's PR comment handling.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>