misc-helper set-app-token: optionally add get more tokens
Originally, the idea was to develop GitGitGadget to see whether it
works, whether it helps, and eventually just move things to
https://github.com/git/git. That is the reason why in many places,
GitGitGadget stores _full_ Pull Request URLs instead of just numbers.

Somewhere along the line, this developer became less sure about this
course of action and instead focused on supporting
https://github.com/gitgitgadget/git well, so in some code paths, that
repository owner "gitgitgadget" is hard-coded.

One of the reasons for this trepidation was the fact that there seemed
to be consensus that GitGitGadget should never be allowed write access
to the git/git repository. It does need to push the tagged iterations,
though, to be able to generate range-diffs between patch series
iterations.

When the idea arose that the tags could still be pushed to
https://github.com/gitgitgadget/git (with a different prefix than the
current tags, i.e. something like `pr-git-<number>` instead of
`pr-<number>`), the plan to support git/git natively picked up steam
again, but there was still the permission problem.

The solution is to have _two_ GitHub Apps: one with write permission to
the gitgitgadget/git repository (and this App is intended to be
installed _only_ on that repository), and a newly-created one:
GitGitGadget-Git (which is intended to be installed on the git/git
repository, and has only write permission on the PR comments and
labels).

With this commit, GitGitGadget learns to pick up the credentials for
installations of that new App, e.g. `git`, and generate a token to do
its work.

The idea is that the private key is available under the Git config key
`gitgitgadget-git.privateKey` (typically specified via the command-line,
in an Azure Pipeline via a secret variable), and the resulting token is
stored under `gitgitgadget.<owner>.githubToken` where `<owner>`
corresponds to the repository owner (i.e. the repository URL is
https://github.com/<owner>/git).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>