misc-helper handle-new-mails: prepare for git/git
To allow GitGitGadget to be used with PRs other than at
https://github.com/gitgitgadget/git, i.e. to support
https://github.com/git/git, let's teach this `misc-helper` command about
(optionally) using a different repository.

The Azure Pipeline "Mirror Git List to GitGitGadget's PRs" at
https://dev.azure.com/gitgitgadget/git/_build?definitionId=5 uses this
command, so we want to be careful not to break things (such as changing
the command-line syntax).

Technically, we do not actually prepare _this command_ for git/git, but
the underlying methods that already accept full Pull Request URLs (which
already contain all the information necessary).

This patch assumes that the config setting `gitgitgadget.githubToken`
has the value necessary for https://github.com/gitgitgadget/git, i.e.
the token generated for the GitHub App "GitGitGadget", and this patch
assumes that for every Pull Request URL thrown at it that has the form
https://github.com/<owner>/git/pull/<number> for owners other than
"gitgitgadget", there exists a corresponding config setting
`gitgitgadget.<owner>.githubToken`, i.e. tokens generated for the GitHub
App "GitGitGadget-Git".

The reason for this split between the owner "gitgitgadget" and all
others is that the GitHub App "GitGitGadget" really needs full write
permissions on the repository (the gitgitgadget/git one, to push tags),
while the GitHub App "GitGitGadget-Git" only needs permissions to add PR
comments and edit PR labels.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>