misc-helper get-pr-meta/get-pr-commits/handle-pr: prepare for git/git
To allow GitGitGadget to be used with PRs other than at
https://github.com/gitgitgadget/git, i.e. to support
https://github.com/git/git, let's teach these `misc-helper` commands
about (optionally) using a different repository.

This is a companion patch to the previous two, but unlike them, these
commands are not wired up in any active Azure Pipeline. Therefore, we
technically do not have to be careful to make the repository owner
optional. We still do because why not.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>