misc-helper handle-pr-*: prepare for git/git
To allow GitGitGadget to be used with PRs other than at
https://github.com/gitgitgadget/git, i.e. to support
https://github.com/git/git, let's teach the `misc-helper` script about
(optionally) using a different repository.

It has to be optional so to keep the Azure Pipeline "GitGitGadget PR
Handler" at https://dev.azure.com/gitgitgadget/git/_build?definitionId=3
working. This pipeline services gitgitgadget/git and needs to keep
running both with this patch and without it because we will not merge
this patch to `master` immediately.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>