MailArchiveHelper: map latestRevision from public-inbox -> lore if necessary
During the transition time when we switch from using a mirror of
https://public-inbox.org/git for "reading the Git mailing list" to a
mirror of https://lore.kernel.org/git/0, there will be a point when
GitGitGadget's record still references a commit in the former repository
while we want to access the latter.

To help with that, assume that during the transition period there will
still be a mirror of the former repository, and an environment variable
`PUBLIC_INBOX_DIR` pointing to it. Then, find out the Message-ID of the
commit, and find the corresponding commit in the new repository.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>