Support alternate config dir for GitGitGadget
Positioning change for tests which need to provide a separate
config to GitGitGadget.get().  Currently, the current directory
is passed.  This will remain the default but user testing should
not need to update the local repo config.  This change will allow
tests to provide an alternate location of the config which
GitGitGadget uses to obtain SMTP information.

Signed-off-by: Chris. Webster <chris@webstech.net>