Improve block quote formatting
The fromString function from "html-to-text" does "lazy"
quoting with only a leading > for the quote.  This makes
cover letters that include quoting hard to read.

This change adds a blockquote formatter that will put `> ` at
the start of each line of the quote.  Quoting inside of quotes
will only add a `>`.  Not perfect but better than relying on
mail formatters to handle lazy quoting.

Signed-off-by: Chris. Webster <chris@webstech.net>