Fix end-of-line handling in the PR description footer parsing
When asking for the PR description, GitHub produces it with CR/LF line
endings. This causes two different problems for the current
implementation of `parsePullRequestDescription()`.

Fix both problems by replacing `\r\n` with `\n` early.

Problem 1
---------

The second regex `/^([-A-Za-z]+:) (.*)\n?$/` did not match lines that
end with `\r\n`, because `(.*)` is not allowed to match `\r`. `\n?` was
a mistake where `\r?` was meant.

As a result, only the last footer was processed, and only if it wasn't
ended with a newline.

Problem 2
---------

The first regex `/^([^]+)\n\n([^]+)$/` wants to split footer on `\n\n`
while it should in fact split on `\r\n\r\n`.

As a result, the split didn't work. Still, it sometimes produced
expected results because it instead split on the title-body boundary,
considering the entire body as footers.

This "footer anywhere" behavior is not preserved, because it is against
the code's intention, and it was considered undesired in discussion for
this patch. For example, consider commit message for git @ 831a488 :

	In some cases it is useful to add additional information after
	the email address on the Cc: footer in a commit log, for instance:

	"Cc: Stable kernel <stable@vger.kernel.org> #v3.4 v3.5 v3.6"

	However, git-send-email refuses to pick up such an invalid
	address when the Email::Valid perl module is available, or just
	uses the whole line as the email address.

In this commit message, there are two Cc: occurrences. Luckily, both of
them are not at the start of the line: one is in the middle and the
other is preceded by quote. Hence they are not mistaken for `Cc:`
footers by GitGitGadget. However, it is easy to imagine that either of
them could have been at the beginning of the line, causing undesired
parsing results.