Handle multiple entries in the same Cc: line again
This got broken somewhere along the lines, most likely when we started
to use `encodeWords()` to quote names when necessary.

This addresses https://github.com/gitgitgadget/gitgitgadget/issues/171

While at it, also support semicolon-separated email addresses, as a
consequence of using the `addressparser()` function that provides us
with this feature "for free".

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>