ci: run the SMTP tests based on ethereal.email
There is a wonderful service at https://ethereal.email/ intended to
support things like Continuous Testing of applications using nodemailer,
which we do.

We already configured the (partially secret) variables in our CI/PR
build, so let's use those settings by copying them into the Git config.

Note: secret variables are not available in PR builds originating from
forks. This is a security measure to prevent users from obtaining the
values of those secret variables. For that reason, we make this new step
conditional: if the secret variable expands to an empty value, we skip
it, otherwise we copy the settings to the Git config.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>