test-lib.ts: do not shadow the global `path` variable
Technically, this patch is not necessary for the current patch series
(whose goal is to fold in the PR description into single patches instead
of sending a full cover letter for one-patch "series"). But it fixes a
problem pointed out by TSLint, so it is better to have it than not to
have it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>