Prepare for formatting the cover letter to variable widths
In preparation for folding in the cover letter of a single-patch
"series" into the only patch (between the `---` line and the diffstat),
we need to reintroduce the method parameter to specify the line length
to use for wrapping the cover letter.

In addition to the width, we also allow configuring a variable
indentation.

This partially reverts 466ec19 (Extract new 'parsePullRequest()',
2019-12-10).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>