Do not omit From: line for single-patch contributions
When teaching GitGitGadget to fold in the cover letter when sending but
a single patch, this developer overlooked that no From: line was
inserted. This means that GitGitGadget's email is used for the
contribution if applied via `git am`.

That happened here:
https://public-inbox.org/git/pull.690.git.git.1578576634678.gitgitgadget@gmail.com/
(it made it into git.git as
https://github.com/git/git/commit/49e268e23e313e8c6b009cd8ebbbaae4316cf6cc)

Let's make sure that those `From:` lines with the original author are
inserted also for single patches.

Note that we _still_ use the sender name in the `From:` header: if Alice
sends a PR on behalf of Bob, i.e. Bob wrote a patch and Alice opens the
PR to send it to the Git mailing list, the sender _is_ Alice, even if
the author _is_ Bob. Therefore, we leave the condition in the
`onBehalfOf =` line alone.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>