GitHubGlue.getPRCommits(): do not expect valid login
It is possible to push commits to GitHub with author/committer
information that does not correspond to any GitHub login. In this case,
the GitHub API apparently does not bother to fill the
`author`/`committer` attributes when returning information for those
commits.

GitGitGadget needs to do better in this case than failing with an
exception like this:

Caught error TypeError: Cannot read property 'login' of null:
TypeError: Cannot read property 'login' of null
    at /home/gitgitgadget-agent/agent/_work/6/s/build/lib/github-glue.js:266:45
    at Array.map (<anonymous>)
    at GitHubGlue.<anonymous> (/home/gitgitgadget-agent/agent/_work/6/s/build/lib/github-glue.js:256:27)
    at Generator.next (<anonymous>) at fulfilled (/home/gitgitgadget-agent/agent/_work/6/s/build/lib/github-glue.js:4:58)
    at processTicksAndRejections (internal/process/task_queues.js:94:5)

Let's just substitute the `login` information (which is what we are
_really_ interested in) by a bogus login.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>