processMails: warn about unfinished buffers, if any
We parse the output of `git log -p` (for performance reasons, we do not
spawn potentially hundreds of `git cat-file` processes). While doing
that, we reconstruct the post-image of the diff, i.e. the new mail,
parsing the number of lines from the diff hunk header.

Just in case that the code is incorrect and fails to reconstruct the
buffer fully, warn about it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>