processMails: adjust for lore.kernel.org
While public-inbox.org only ever adds the new mails, lore.kernel.org has
a new layout: it has but a single file in every commit's tree, which is
the current mail. Therefore, the diff has plenty of `-` lines because it
"removes" the previous mail (it is still in the history, of course, and
therefore in the Git object database).

But when we parse the log, we still expect the old layout, where only
`+` lines appear in the diffs.

Let's adjust the code so it can handle lore.kernel.org's layout.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>