processMails: make sure to get complete mails
When parsing the mails from lore.kernel.org's layout, we actually only
get the diffs between the previously-current and the now-current mail.
Since there are matching lines (such as the empty line between the mail
header and the body), we have plenty of context lines, and we could
potentially end up with incomplete post-images because the diff simply
might not contain all of the lines.

Let's increase the diff context to an insanely large number to ensure
that we _do_ get the entire post-image.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>