Add GitHubGlue tests
This test requires setup.  It will run successfully if setup has
not been done.  If the test fails, there may be a pull request and a
branch to be deleted on github and the local repo.  The test will
attempt cleanup of a previously failed test.

Setup:
gitgitgadget.githubtest.githubuser must be configured for this test to
identify your GitHub login.
The value must also be configured to identify a test repo to use as
gitgitgadget.<login>.githubrepo.

Additionally, a GitHub personal access token must be set for the
login to succeed.  This can be restricted to the test repo.

For example, these configuration settings are needed (where
`octokity` is your GitHub login):
git config --add gitgitgadget.githubtest.githubuser octokity
git config --add gitgitgadget.octokity.githubrepo ggg-test
git config --add gitgitgadget.octokity.githubtoken feedbeef...

The test repo must exist in github and locally.  It is expected to be
located at the same directory level as this project (ie ../).

Signed-off-by: Chris. Webster <chris@webstech.net>