Allow empty PR description for single-patch contributions
Back when we _always_ sent a cover letter, it made sense to ask for a
non-empty description. However, since we started "folding in the cover
letter" for single-patch contributions, that restriction wants to be
loosened: no need for a description if we're only sending one mail to
begin with.

Example: https://github.com/git/git/pull/740

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>