Update dugite to the latest version
This also upgrades the embedded Git version from v2.20.1 to v2.23.1,
which fixes an important bug: when a `git push` fails where the
credentials are embedded in the push URL, that URL will be displayed
verbatim, including credentials.

In GitGitGadget's context, this means that the token used by
GitGitGadget for pushing might leak into a PR comment. Not good, even if
that token is only valid for an hour.

So let's upgrade and be safe.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>