CIHelper: _never_ leak the token in the comments
Recently, PRs were opened that try to edit the GitHub workflow. This is
apparently not permitted for apps such as GitGitGadget, so when we try
to push the tag with a modified GitHub workflow, GitHub refuses.

Unfortunately, Git has a bug where it fails to anonymize the URL, and
since we embed the token in the URL so that we can push, that token is
leaked.

Always redact it before showing the user the error message.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>