Decode Author Name on Inserted From Line
Git encodes the author name on patch emails.  The encoded name may
have a newline between the name and the email address.  The regex for
`authorMatch` now allows a newline to be part of the `From:` header.
The name will now be decoded when it is added to the email body
as a `From:` line.

Signed-off-by: Chris. Webster <chris@webstech.net>