Handle missing secrets in github push workflow
Git does not allow leading spaces in the `GIT_CONFIG_PARAMETERS`
environment variable.  This corrects that for when no secrets are
defined for the workflow.

Additionally, the workflow will now get an error if only some of
the secrets in a group (there are two) are defined.

The format of the smtp related secrets are now just the basic
values and not the internal formats.

Signed-off-by: Chris. Webster <chris@webstech.net>