tools: new script to clean up GitHubGlue test artifacts
delete-test-branches.ts processes the command line and calls
delete-branches in delete-ci-test-branches.ts.

delete-ci-test-branches.ts retrieves a list of branches from GitHub
and deletes the ones older than the specified time.  The deletion
of the refs will result in any open pull requests on the branch
being closed.  This routine is also intended to be used from a
GitHub workflow.

Signed-off-by: Chris. Webster <chris@webstech.net>